Initial commit onto develop with band create scripts
